Add NaCl even hkl to ImageCalibrants.py Fix bug in polygon mask in GSASIIimage.py Fix copy bugs in GSASIIimgGUI.py Make float image for float tiff file in GSASIIIO.py
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@251 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ GSASIIIO.py
@@ -694,7 +694,7 @@ def GetTifData(filename,imageOnly=False):
                 if not imageOnly:
                     print 'Read APS PE-detector tiff file: ',filename
                 if dataType == 5:
-                    image = np.array(ar.array('f',File.read(4*Npix)),dtype=np.int32)
+                    image = np.array(ar.array('f',File.read(4*Npix)),dtype=np.float32)
                 else:
                     image = np.array(ar.array('I',File.read(4*Npix)),dtype=np.int32)
         elif IFD[273][2][0] == 4096:
